feat: warn when restart monitoring is unexpectedly delayed
When lots of files match the spec for files to be monitored, it
can take a long time for chokidar to load them all. While this
happens, nodemon will not restart on file change, but there is
no indication of that. Add a warning message if monitor scanning
is taking a long time, and then periodic repeats beyond that, and
a confirmation message once monitor scanning is complete and
nodemon is actually watching for change.